Move time constants to time module
